Refresh data in EditTeamView now correctly updates tickets from server

+++ app/src/main/java/com/normorovers/mmt/app/event/mmtevent/api/Teams.kt
@@ -9,12 +9,12 @@ interface Teams {
     fun getTeams(): Call<List<Team>>
 
     @GET("teams/{id}")
-    fun getTeam(@Path("id") id: Int): Call<Team>
+	fun getTeam(@Path("id") id: Long): Call<Team>
 
     @GET("teams/{id}")
     fun getTeam(
-            @Path("id") id: Int,
-            @Query("load_tickets") loadTickets: Boolean
+			@Path("id") id: Long,
+			@Query("load_tickets") loadTickets: Boolean
     ): Call<Team>
 
     @GET("teams/{id}/point_logs")
